Frontend Add

+++ app/Http/Controllers/SubCatagoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Catagory;
+use App\Models\Product;
 use App\Models\Subcatagory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -16,7 +17,9 @@ class SubCatagoryController extends Controller
      */
     public function index()
     {
-        $sub_catagories = Subcatagory::with('Catagory')->select('id', 'catagory_id', 'subcatagory_name', 'created_at')->latest('id')->get();
+        $sub_catagories = Subcatagory::with('Catagory')
+            ->select('id', 'catagory_id', 'subcatagory_name', 'created_at')
+            ->latest('id')->get();
         return view('backend.subcatagory.index', [
             'subcatagories' => $sub_catagories,
         ]);
@@ -117,16 +120,26 @@ class SubCatagoryController extends Controller
      */
     public function destroy($id)
     {
-        Subcatagory::findorfail($id)->delete();
-        return back()->with('delete', 'Subcatagory Deleted Successfully');
+        $product = Product::where('subcatagory_id', $id)->get()->count();
+        if ($product > 0) {
+            return back()->with('warning', 'Theres a Product Under This Subcatagory');
+        } else {
+            Subcatagory::findorfail($id)->delete();
+            return back()->with('delete', 'Subcatagory Deleted Successfully');
+        }
     }
     public function MarkdeleteSubCatagory(Request $request)
     {
         if ($request->filled('delete')) {
-           foreach ($request->delete as $value) {
-            Subcatagory::findorfail($value)->delete();
-           }
-           return back()->with('delete', 'Subcatagory Deleted Successfully');
+            foreach ($request->delete as $value) {
+                $product = Product::where('subcatagory_id', $value)->get()->count();
+                if ($product > 0) {
+                    return back()->with('warning', 'Theres a Product Under This Subcatagory');
+                } else {
+                    Subcatagory::findorfail($value)->delete();
+                }
+            }
+            return back()->with('delete', 'Subcatagory Deleted Successfully');
         } else {
             return back()->with('warning', 'No Item Selected');
         }
